requirements.txt
beautifulsoup4==4.7.1
cycler==0.10.0
dataclasses==0.6
django-graphos==0.3.41
gcloud==0.17.0
googleapis-common-protos==1.5.5
httplib2==0.12.0
jws==0.1.3
matplotlib==1.5.3
numpy==1.16.0
oauth2client==3.0.0
pandas==0.19.1
patsy==0.5.1
Pillow==5.4.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
pymongo==3.7.2
pyparsing==2.3.1
Pyrebase==3.0.27
python-dateutil==2.7.5
python-jwt==2.0.1
pytz==2018.7
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
soupsieve==1.7.2
sqlparse==0.2.4
statsmodels==0.9.0
virtualenv==16.2.0
zerosms==0.0.1